Move HabitColors to habits domain model
