+++ feature/habits/domain/src/commonMain/kotlin/space/be1ski/vibits/feature/habits/domain/model/HabitConfig.kt
@@ -1,9 +1,4 @@
-@file:Suppress("MagicNumber")
-
 package space.be1ski.vibits.feature.habits.domain.model
-
-/** Material Green 500 — default color when none is specified in config. */
-const val DEFAULT_HABIT_COLOR: Long = 0xFF4CAF50L
 
 /**
  * Habit configuration entry.
@@ -14,5 +9,5 @@ data class HabitConfig(
   /** User-friendly label. */
   val label: String,
   /** Habit color as ARGB Long. */
-  val color: Long = DEFAULT_HABIT_COLOR,
+  val color: Long = DefaultHabitColor,
 )
