+++ feature/memos/data/build.gradle.kts
@@ -17,7 +17,6 @@ kotlin {
     commonMain {
       dependencies {
         implementation(projects.core.platform)
-        implementation(projects.core.ui)
         implementation(projects.core.utils)
         implementation(projects.feature.auth.domain)
         implementation(projects.feature.habits.domain)
