+++ feature/onboarding/domain/src/commonTest/kotlin/space/be1ski/vibits/feature/onboarding/domain/usecase/OnboardingUseCasesTest.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.test.runTest
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.atStartOfDayIn
-import space.be1ski.vibits.feature.habits.domain.model.DEFAULT_HABIT_COLOR
+import space.be1ski.vibits.feature.habits.domain.model.DefaultHabitColor
 import space.be1ski.vibits.feature.memos.data.offline.OfflineMemoDto
 import space.be1ski.vibits.feature.memos.data.offline.OfflineMemosFileDto
 import space.be1ski.vibits.feature.memos.data.test.FakeOfflineMemoStorage
@@ -82,7 +82,7 @@ class OnboardingUseCasesTest {
       val createMemo = CreateMemoUseCase(memosRepository)
       val useCase = CreateFirstHabitUseCase(createMemo)
 
-      val result = useCase("Morning Exercise", "custom", DEFAULT_HABIT_COLOR)
+      val result = useCase("Morning Exercise", "custom", DefaultHabitColor)
 
       assertTrue(result.isSuccess)
       assertEquals(1, memosRepository.createMemoCalls)
@@ -163,7 +163,7 @@ class OnboardingUseCasesTest {
       val createMemo = CreateMemoUseCase(memosRepository)
       val useCase = CreateFirstHabitUseCase(createMemo)
 
-      val result = useCase("Exercise", "custom", DEFAULT_HABIT_COLOR)
+      val result = useCase("Exercise", "custom", DefaultHabitColor)
 
       assertTrue(result.isFailure)
       assertEquals(1, memosRepository.createMemoCalls)
@@ -179,7 +179,7 @@ class OnboardingUseCasesTest {
       val createMemo = CreateMemoUseCase(memosRepository)
       val useCase = CreateFirstHabitUseCase(createMemo)
 
-      val result = useCase("Read Every Day", "read", DEFAULT_HABIT_COLOR)
+      val result = useCase("Read Every Day", "read", DefaultHabitColor)
 
       assertTrue(result.isSuccess)
       val createdContent = memosRepository.lastCreatedContent
