+++ feature/onboarding/presentation/src/commonMain/kotlin/space/be1ski/vibits/feature/onboarding/presentation/state/OnboardingState.kt
@@ -1,6 +1,6 @@
 package space.be1ski.vibits.feature.onboarding.presentation.state
 
-import space.be1ski.vibits.feature.habits.domain.model.DEFAULT_HABIT_COLOR
+import space.be1ski.vibits.feature.habits.domain.model.DefaultHabitColor
 import space.be1ski.vibits.feature.onboarding.domain.model.HabitPreset
 
 data class OnboardingState(
@@ -9,7 +9,7 @@ data class OnboardingState(
   val selectedPresetId: String? = null,
   val selectedPresetName: String? = null,
   val habitName: String = "",
-  val selectedColor: Long = DEFAULT_HABIT_COLOR,
+  val selectedColor: Long = DefaultHabitColor,
   val isCreatingHabit: Boolean = false,
   val creationError: String? = null,
   val habitCreated: Boolean = false,
